+++ contracts/LockedDeal/LockedPoolz.sol
@@ -8,12 +8,22 @@ contract LockedPoolz is Manageable {
     constructor() public {
         Index = 0;
     }
-    
+
     // add contract name
     string public name;
 
-    event NewPoolCreated(uint256 PoolId, address Token, uint64 FinishTime, uint256 StartAmount, address Owner);
-    event PoolOwnershipTransfered(uint256 PoolId, address NewOwner, address OldOwner);
+    event NewPoolCreated(
+        uint256 PoolId,
+        address Token,
+        uint64 FinishTime,
+        uint256 StartAmount,
+        address Owner
+    );
+    event PoolOwnershipTransfered(
+        uint256 PoolId,
+        address NewOwner,
+        address OldOwner
+    );
     event PoolApproval(uint256 PoolId, address Spender, uint256 Amount);
 
     struct Pool {
@@ -21,7 +31,7 @@ contract LockedPoolz is Manageable {
         uint256 Amount;
         address Owner;
         address Token;
-        mapping(address => uint) Allowance;
+        mapping(address => uint256) Allowance;
     }
     // transfer ownership
     // allowance
@@ -31,52 +41,67 @@ contract LockedPoolz is Manageable {
     mapping(address => uint256[]) MyPoolz;
     uint256 internal Index;
 
-    modifier isTokenValid(address _Token){
+    modifier isTokenValid(address _Token) {
         require(isTokenWhiteListed(_Token), "Need Valid ERC20 Token"); //check if _Token is ERC20
         _;
     }
 
-    modifier isPoolValid(uint256 _PoolId){
+    modifier isPoolValid(uint256 _PoolId) {
         require(_PoolId < Index, "Pool does not exist");
         _;
     }
 
-    modifier isPoolOwner(uint256 _PoolId){
-        require(AllPoolz[_PoolId].Owner == msg.sender, "You are not Pool Owner");
+    modifier isPoolOwner(uint256 _PoolId) {
+        require(
+            AllPoolz[_PoolId].Owner == msg.sender,
+            "You are not Pool Owner"
+        );
         _;
     }
 
-    modifier isAllowed(uint256 _PoolId, uint256 _amount){
-        require(_amount <= AllPoolz[_PoolId].Allowance[msg.sender], "Not enough Allowance");
+    modifier isAllowed(uint256 _PoolId, uint256 _amount) {
+        require(
+            _amount <= AllPoolz[_PoolId].Allowance[msg.sender],
+            "Not enough Allowance"
+        );
         _;
     }
 
-    modifier isLocked(uint256 _PoolId){
+    modifier isLocked(uint256 _PoolId) {
         require(AllPoolz[_PoolId].UnlockTime > now, "Pool is Unlocked");
         _;
     }
 
-    modifier notZeroAddress(address _address){
+    modifier notZeroAddress(address _address) {
         require(_address != address(0x0), "Zero Address is not allowed");
         _;
     }
 
-    modifier isGreaterThanZero(uint256 _num){
+    modifier isGreaterThanZero(uint256 _num) {
         require(_num > 0, "Array length should be greater than zero");
         _;
     }
 
-    modifier isBelowLimit(uint256 _num){
+    modifier isBelowLimit(uint256 _num) {
         require(_num <= maxTransactionLimit, "Max array length limit exceeded");
         _;
     }
 
-    function SplitPool(uint256 _PoolId, uint256 _NewAmount , address _NewOwner) internal returns(uint256) {
+    function SplitPool(
+        uint256 _PoolId,
+        uint256 _NewAmount,
+        address _NewOwner
+    ) internal returns (uint256) {
         Pool storage pool = AllPoolz[_PoolId];
         require(pool.Amount >= _NewAmount, "Not Enough Amount Balance");
         uint256 poolAmount = SafeMath.sub(pool.Amount, _NewAmount);
         pool.Amount = poolAmount;
-        uint256 poolId = CreatePool(pool.Token, pool.UnlockTime, _NewAmount, _NewOwner);
+        uint256 poolId = CreatePool(
+            pool.Token,
+            pool.UnlockTime,
+            _NewAmount,
+            _NewOwner
+        );
         return poolId;
     }
 
@@ -86,7 +111,7 @@ contract LockedPoolz is Manageable {
         uint64 _FinishTime, //Until what time the pool will work
         uint256 _StartAmount, //Total amount of the tokens to sell in the pool
         address _Owner // Who the tokens belong to
-    ) internal returns(uint256){
+    ) internal returns (uint256) {
         //register the pool
         AllPoolz[Index] = Pool(_FinishTime, _StartAmount, _Owner, _Token);
         MyPoolz[_Owner].push(Index);
