+++ contracts/LockedDealV2/LockedControl.sol
@@ -34,7 +34,13 @@ contract LockedControl is LockedPoolz {
         uint256 _PoolId,
         uint256 _NewAmount,
         address _NewOwner
-    ) external isPoolValid(_PoolId) isPoolOwner(_PoolId) isLocked(_PoolId) returns(uint256) {
+    )
+        external
+        isPoolValid(_PoolId)
+        isPoolOwner(_PoolId)
+        isLocked(_PoolId)
+        returns (uint256)
+    {
         uint256 poolId = SplitPool(_PoolId, _NewAmount, _NewOwner);
         return poolId;
     }
@@ -43,12 +49,23 @@ contract LockedControl is LockedPoolz {
         uint256 _PoolId,
         uint256 _Amount,
         address _Spender
-    ) external isPoolValid(_PoolId) isPoolOwner(_PoolId) isLocked(_PoolId) notZeroAddress(_Spender) {
+    )
+        external
+        isPoolValid(_PoolId)
+        isPoolOwner(_PoolId)
+        isLocked(_PoolId)
+        notZeroAddress(_Spender)
+    {
         Allowance[_PoolId][_Spender] = _Amount;
         emit PoolApproval(_PoolId, _Spender, _Amount);
     }
 
-    function GetPoolAllowance(uint256 _PoolId, address _Address) public view isPoolValid(_PoolId) returns(uint256){
+    function GetPoolAllowance(uint256 _PoolId, address _Address)
+        public
+        view
+        isPoolValid(_PoolId)
+        returns (uint256)
+    {
         return Allowance[_PoolId][_Address];
     }
 
@@ -56,10 +73,16 @@ contract LockedControl is LockedPoolz {
         uint256 _PoolId,
         uint256 _Amount,
         address _Address
-    ) external isPoolValid(_PoolId) isAllowed(_PoolId, _Amount) isLocked(_PoolId) returns(uint256) {
+    )
+        external
+        isPoolValid(_PoolId)
+        isAllowed(_PoolId, _Amount)
+        isLocked(_PoolId)
+        returns (uint256)
+    {
         uint256 poolId = SplitPool(_PoolId, _Amount, _Address);
         uint256 _NewAmount = Allowance[_PoolId][msg.sender] - _Amount;
-        Allowance[_PoolId][msg.sender]  = _NewAmount;
+        Allowance[_PoolId][msg.sender] = _NewAmount;
         return poolId;
     }
 
@@ -69,9 +92,12 @@ contract LockedControl is LockedPoolz {
         uint256 _FinishTime, //Until what time the pool will end
         uint256 _StartAmount, //Total amount of the tokens to sell in the pool
         address _Owner // Who the tokens belong to
-    ) external payable notZeroAddress(_Owner) returns(uint256) {
+    ) external payable notZeroAddress(_Owner) returns (uint256) {
         TransferInToken(_Token, msg.sender, _StartAmount);
-        if(WhiteList_Address != address(0) && !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))){
+        if (
+            WhiteList_Address != address(0) &&
+            !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))
+        ) {
             PayFee(Fee);
         }
         CreatePool(_Token, _StartTime, _FinishTime, _StartAmount, _Owner);
@@ -83,7 +109,9 @@ contract LockedControl is LockedPoolz {
         uint256[] calldata _FinishTime,
         uint256[] calldata _StartAmount,
         address[] calldata _Owner
-    )   external payable
+    )
+        external
+        payable
         isGreaterThanZero(_Owner.length)
         isBelowLimit(_Owner.length)
     {
@@ -91,12 +119,21 @@ contract LockedControl is LockedPoolz {
         require(_StartTime.length == _FinishTime.length, "Date Array Invalid");
         require(_Owner.length == _StartAmount.length, "Amount Array Invalid");
         TransferInToken(_Token, msg.sender, Array.getArraySum(_StartAmount));
-        if(WhiteList_Address != address(0) && !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))){
+        if (
+            WhiteList_Address != address(0) &&
+            !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))
+        ) {
             PayFee(Fee * _Owner.length);
         }
         uint256 firstPoolId = Index;
-        for(uint i=0 ; i < _Owner.length; i++){
-            CreatePool(_Token, _StartTime[i], _FinishTime[i], _StartAmount[i], _Owner[i]);
+        for (uint256 i = 0; i < _Owner.length; i++) {
+            CreatePool(
+                _Token,
+                _StartTime[i],
+                _FinishTime[i],
+                _StartAmount[i],
+                _Owner[i]
+            );
         }
         uint256 lastPoolId = Index - 1;
         emit MassPoolsCreated(firstPoolId, lastPoolId);
@@ -109,20 +146,35 @@ contract LockedControl is LockedPoolz {
         uint256[] calldata _FinishTime,
         uint256[] calldata _StartAmount,
         address[] calldata _Owner
-    )   external payable
+    )
+        external
+        payable
         isGreaterThanZero(_StartTime.length)
         isBelowLimit(_Owner.length * _FinishTime.length)
     {
         require(_Owner.length == _StartAmount.length, "Amount Array Invalid");
         require(_FinishTime.length == _StartTime.length, "Date Array Invalid");
-        TransferInToken(_Token, msg.sender, Array.getArraySum(_StartAmount) * _FinishTime.length);
+        TransferInToken(
+            _Token,
+            msg.sender,
+            Array.getArraySum(_StartAmount) * _FinishTime.length
+        );
         uint256 firstPoolId = Index;
-        if(WhiteList_Address != address(0) && !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))){
+        if (
+            WhiteList_Address != address(0) &&
+            !(isUserWithoutFee(msg.sender) || isTokenWithoutFee(_Token))
+        ) {
             PayFee(Fee * _Owner.length * _FinishTime.length);
         }
-        for(uint i=0 ; i < _FinishTime.length ; i++){
-            for(uint j=0 ; j < _Owner.length ; j++){
-                CreatePool(_Token, _StartTime[i], _FinishTime[i], _StartAmount[j], _Owner[j]);
+        for (uint256 i = 0; i < _FinishTime.length; i++) {
+            for (uint256 j = 0; j < _Owner.length; j++) {
+                CreatePool(
+                    _Token,
+                    _StartTime[i],
+                    _FinishTime[i],
+                    _StartAmount[j],
+                    _Owner[j]
+                );
             }
         }
         uint256 lastPoolId = Index - 1;
